# PERF: Improve JS app boot speed by optimizing `customResolve()`