@davidtaylorhq 2021-11-17
I'm not sure why this is consistently failing the Firefox tests. It seems to fail a different test each time, and everything works fine in firefox locally. Will do some digging 👀